requirements.txt
Flask==3.0.0
firebase-admin==6.2.0
google-cloud-firestore==2.14.0
gunicorn==21.2.0
Werkzeug==3.0.1